+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.9.0
 
 * Adding the `phonetics/alpha-sis` namespace.
+* Adding the `phonetics/fuzzy-soundex` namespace.
 * Adding the `stemmers/uea-lite` namespace.
 
 ## 0.8.0